+++ app/scripts/components/frontpage.ts
@@ -16,14 +16,14 @@ export default angular.module("korpApp").component("frontpage", {
                 >
                     <div
                         ng-if="$root._settings['description']"
-                        ng-bind-html="$root._settings['description'] | locObj:lang | trust"
+                        ng-bind-html="$root._settings['description'] | locObj:$root.lang | trust"
                     ></div>
 
                     <div ng-if="$root._settings['mode_description']">
                         <h3 ng-if="$root._settings['mode']['label']" class="font-bold">
-                            {{$root._settings['mode']['label'] | locObj:lang}}
+                            {{$root._settings['mode']['label'] | locObj:$root.lang}}
                         </h3>
-                        <div ng-bind-html="$root._settings['mode_description'] | locObj:lang | trust"></div>
+                        <div ng-bind-html="$root._settings['mode_description'] | locObj:$root.lang | trust"></div>
                     </div>
                 </section>